+++ python/__init__.py
@@ -66,8 +66,6 @@ from .check_tt64_crc import check_tt64_crc
 
 from .append_crc32c import append_crc32c
 
-from .strip_ax25_header import strip_ax25_header
-
 from .snet_deframer import snet_deframer
 
 from .beesat_classifier import beesat_classifier
@@ -78,7 +76,6 @@ from .k2sat_deframer import k2sat_deframer
 
 from .cc11xx_packet_crop import cc11xx_packet_crop
 from .check_cc11xx_crc import check_cc11xx_crc
-from .cc11xx_remove_length import cc11xx_remove_length
 
 from .pwsat2_telemetry_parser import pwsat2_telemetry_parser
 from .pwsat2_submitter import pwsat2_submitter
@@ -98,8 +95,6 @@ from .check_swiatowid_crc import check_swiatowid_crc
 from .swiatowid_packet_split import swiatowid_packet_split
 
 from .manchester_sync import manchester_sync
-
-from .header_remover import header_remover
 
 from .aausat4_remove_fsm import aausat4_remove_fsm